Rename Base to Kernel, move to workable state
The module is renamed to SlmCmfKernel to reflect the function in the CMF
system. Furthermore is the metadata stripped from the kernel as it is more
navigation related and will be placed in the navigation module together
with menu/breadcrumb building etcetera.

Parsing changed to reflect structure of original routing as best as possible.
Additional changes are made to use service manager instead of dependency
injection containers.