Remove Doctrine code from Kernel
The Kernel is constructed with contracts and all concrete classes from the domain layer are removed. The business
logic is presented with Listeners and Parsers, and the kernel requires an additional module to provide the models
for its parsing. Currently a second module SlmCmfKernelDoctrineOrm is created to supply models for the Doctrine
ORM. It will be trivial to create same features using Zend\Db or another data access object.